added in example native C library and corresponding Java
This is a complete setup for implementing some native stuff
with nice java glue.

Also checked in is the compiled binary .so file, this way people
not dealing with native code can just stick to their normal
Java-only build procedures.

To build native code, FIRST make sure the java class is compiled,
then run 'make build-headers' to update the corresponding .h files
(if necesary), then build the native library with 'make build-native'
Finally re-build the java project ('make build-debug' if from
the command line) and it will properly link it with the freshly
compiled native .so file